[step2] refactor: utils.js 분리
