+++ src/components/Issue/issueItem.js
@@ -1,7 +1,8 @@
 import {SELECTOR} from "../../const.js";
-import {filterIssueByStatus, getIssueTplStr, pipe} from "../../utils.js";
+import {filterIssueByStatus, getIssueTplStr} from "../../utils/util.js";
 import {issuesAtom, issueStatusAtom} from "../../store/atom.js";
 import {useAtomValue} from "../../store/atomHooks.js";
+import {pipe} from "../../utils/functional.js";
 
 const getIssues = useAtomValue(issuesAtom);
 const getIssueStatus = useAtomValue(issueStatusAtom);
