+++ src/utils/fetch.js
@@ -1,0 +1,3 @@
+export const fetchIssueData = async () => await ((await fetch("../data-sources/issues.json")).json());
+
+export const fetchLabelData = async () => await ((await fetch("../data-sources/labels.json")).json());
